refactor(LottoGame, MatchCategory): move `MatchCategory` enum from `LottoGame` to its own file
Relocate the `MatchCategory` enum to a dedicated file:
- improves code organization by separating concerns
- makes the enum easier to locate and maintain
