+++ src/main/kotlin/lotto/MatchCategory.kt
@@ -1,0 +1,9 @@
+package lotto
+
+enum class MatchCategory {
+    THREE_MATCHES,
+    FOUR_MATCHES,
+    FIVE_MATCHES,
+    FIVE_MATCHES_AND_BONUS,
+    SIX_MATCHES
+}
